Removed various redundant flags
- FlowServiceInfo capabilities removed - no use on the device itself
- FlowServiceInfo stages now set internally instead of via builder
- FlowServiceInfo backgroundOnly removed - can be managed by the service itself
- PaymentServiceInfo operatingMode removed - no clear use case
