+++ FlowServiceSample/src/main/java/com/aevi/sdk/pos/flow/flowservicesample/service/PostCardService.java
@@ -1,15 +1,15 @@
 package com.aevi.sdk.pos.flow.flowservicesample.service;
 
 
-import com.aevi.sdk.pos.flow.flowservicesample.ui.PrePaymentActivity;
+import com.aevi.sdk.pos.flow.flowservicesample.ui.PostCardActivity;
 import com.aevi.sdk.pos.flow.model.TransactionRequest;
-import com.aevi.sdk.pos.flow.service.BasePrePaymentService;
+import com.aevi.sdk.pos.flow.service.BasePostCardService;
 
-public class PostCardService extends BasePrePaymentService {
+public class PostCardService extends BasePostCardService {
 
     @Override
     protected void processRequest(String clientMessageId, TransactionRequest transactionRequest) {
-        launchActivity(PrePaymentActivity.class, clientMessageId, transactionRequest);
+        launchActivity(PostCardActivity.class, clientMessageId, transactionRequest);
     }
 
     @Override
