+++ FlowServiceSample/src/main/java/com/aevi/sdk/pos/flow/flowservicesample/ui/PostCardActivity.java
@@ -1,0 +1,10 @@
+package com.aevi.sdk.pos.flow.flowservicesample.ui;
+
+/**
+ * Sample activity for post-card-reading stage. See base class.
+ *
+ * This sample illustrates the data made available to the flow application at these stages,
+ * and in what different ways a post-card-reading flow app can augment the transaction.
+ */
+public class PostCardActivity extends BasePreProcessingActivity {
+}
