+++ api-constants/src/main/java/com/aevi/sdk/flow/constants/AugmentedDataKeys.java
@@ -1,0 +1,10 @@
+package com.aevi.sdk.flow.constants;
+
+public interface AugmentedDataKeys {
+
+    String AUGMENTED_DATA_AMOUNTS = "amounts";
+    String AUGMENTED_DATA_CURRENCY = "currency";
+    String AUGMENTED_DATA_OPTIONS = "options";
+    String AUGMENTED_DATA_SPLIT = "split";
+    String AUGMENTED_DATA_CANCELLED = "cancelled";
+}
