+++ api-constants/src/main/java/com/aevi/sdk/flow/constants/InternalDataKeys.java
@@ -1,0 +1,6 @@
+package com.aevi.sdk.flow.constants;
+
+public interface InternalDataKeys {
+
+    String BACKGROUND_PROCESSING = "backgroundProcessing";
+}
