+++ api-constants/src/main/java/com/aevi/sdk/flow/constants/RequestSource.java
@@ -3,7 +3,7 @@ package com.aevi.sdk.flow.constants;
 
 public interface RequestSource {
 
-    String AEVI_V2 = "Aevi v2";
-    String AEVI_V3 = "Aevi v3";
+    String AEVI_V2 = "AEVI v2";
+    String AEVI_POS_FLOW = "AEVI POS Flow";
 
 }
