+++ payment-initiation-api/src/main/java/com/aevi/sdk/pos/flow/model/FlowAppInfo.java
@@ -1,7 +1,8 @@
 package com.aevi.sdk.pos.flow.model;
 
+import com.aevi.sdk.pos.flow.PaymentClient;
+
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -9,27 +10,19 @@ import java.util.List;
  */
 public class FlowAppInfo {
 
-    public static final String AUGMENTED_DATA_AMOUNTS = "amounts";
-    public static final String AUGMENTED_DATA_OPTIONS = "options";
-    public static final String AUGMENTED_DATA_SPLIT = "split";
-    public static final String AUGMENTED_DATA_CANCELLED = "cancelled";
-
-    private final String appName;
-    private final String[] capabilities;
+    private final String flowAppId;
     private final String stage;
     private final List<String> augmentedData;
 
     /**
      * Create a new FlowAppInfo instance with augmented data information.
      *
-     * @param appName       The flow application name
-     * @param capabilities  The flow capabilities
+     * @param flowAppId     The flow application id (as per {@link com.aevi.sdk.flow.model.FlowServiceInfo})
      * @param stage         The stage at which the flow app is called
      * @param augmentedData The list of augmented data
      */
-    public FlowAppInfo(String appName, String[] capabilities, String stage, List<String> augmentedData) {
-        this.appName = appName;
-        this.capabilities = capabilities;
+    public FlowAppInfo(String flowAppId, String stage, List<String> augmentedData) {
+        this.flowAppId = flowAppId;
         this.stage = stage;
         this.augmentedData = augmentedData;
     }
@@ -37,32 +30,31 @@ public class FlowAppInfo {
     /**
      * Create a new FlowAppInfo instance without any augmented data.
      *
-     * @param appName      The flow application name
-     * @param capabilities The flow capabilities
-     * @param stage        The stage at which the flow app is called
+     * @param flowAppId The flow application name
+     * @param stage     The stage at which the flow app is called
      */
-    public FlowAppInfo(String appName, String[] capabilities, String stage) {
-        this(appName, capabilities, stage, new ArrayList<String>());
+    public FlowAppInfo(String flowAppId, String stage) {
+        this(flowAppId, stage, new ArrayList<String>());
     }
 
     /**
-     * Get the flow application name.
+     * Get the flow application id, as per {@link com.aevi.sdk.flow.model.FlowServiceInfo}.
      *
-     * @return The flow application name
+     * This can be used to look up further details about the flow application via {@link PaymentClient#getFlowServices()}.
+     *
+     * @return The flow application id
      */
-    public String getAppName() {
-        return appName;
+    public String getFlowAppId() {
+        return flowAppId;
     }
 
     /**
-     * Get the flow application capabilities (such as loyalty, dcc, etc).
+     * Get the stage for which this flow application executed in.
      *
-     * This will help determining for what purposes the transaction was augmented.
-     *
-     * @return The capabilities
+     * @return The flow stage
      */
-    public String[] getCapabilities() {
-        return capabilities;
+    public String getStage() {
+        return stage;
     }
 
     /**
@@ -70,8 +62,7 @@ public class FlowAppInfo {
      *
      * Note that this simply indicates what was augmented - not to what. The {@link PaymentResponse} can be parsed to review that information.
      *
-     * See {@link #AUGMENTED_DATA_AMOUNTS}, {@link #AUGMENTED_DATA_OPTIONS}, {@link #AUGMENTED_DATA_CANCELLED} and {@link #AUGMENTED_DATA_SPLIT} for
-     * the possible options.
+     * See documentation for possible values.
      *
      * @return The list of data that was augmented.
      */
@@ -90,11 +81,7 @@ public class FlowAppInfo {
 
         FlowAppInfo that = (FlowAppInfo) o;
 
-        if (appName != null ? !appName.equals(that.appName) : that.appName != null) {
-            return false;
-        }
-        // Probably incorrect - comparing Object[] arrays with Arrays.equals
-        if (!Arrays.equals(capabilities, that.capabilities)) {
+        if (flowAppId != null ? !flowAppId.equals(that.flowAppId) : that.flowAppId != null) {
             return false;
         }
         if (stage != null ? !stage.equals(that.stage) : that.stage != null) {
@@ -105,8 +92,7 @@ public class FlowAppInfo {
 
     @Override
     public int hashCode() {
-        int result = appName != null ? appName.hashCode() : 0;
-        result = 31 * result + Arrays.hashCode(capabilities);
+        int result = flowAppId != null ? flowAppId.hashCode() : 0;
         result = 31 * result + (stage != null ? stage.hashCode() : 0);
         result = 31 * result + (augmentedData != null ? augmentedData.hashCode() : 0);
         return result;
@@ -115,8 +101,7 @@ public class FlowAppInfo {
     @Override
     public String toString() {
         return "FlowAppInfo{" +
-                "appName='" + appName + '\'' +
-                ", capabilities=" + Arrays.toString(capabilities) +
+                "flowAppId='" + flowAppId + '\'' +
                 ", stage=" + stage +
                 ", augmentedData=" + augmentedData +
                 '}';
